Migrate off postgres functions and on to the diesel ORM
Move the remaining API calls to postgres functions to leverage diesel directly
Leave the postgres functions in place for now so we don't break jobsrv

+++ components/builder-api/src/server/authorize.rs
@@ -99,5 +99,5 @@ pub fn check_origin_member(
 ) -> Result<bool> {
     let conn = req.state().db.get_conn().map_err(Error::DbError)?;
 
-    Origin::check_membership(origin, account_id, &*conn).map_err(Error::DieselError)
+    Origin::check_membership(origin, account_id as i64, &*conn).map_err(Error::DieselError)
 }
